update pause game menu

+++ Assets/Scripts/Settings.cs
@@ -5,27 +5,27 @@ using UnityEngine;
 [System.Serializable]
 public class Settings
 {
-    public int sound;
-    public int sfx;
-    public int mouseSensitivity;
+    public float sound;
+    public float sfx;
+    public float mouseSensitivity;
 
-    public Settings(int sound, int sfx, int mouseSensitivity)
+    public Settings(float sound, float sfx, float mouseSensitivity)
     {
         this.sound = sound;
         this.sfx = sfx;
         this.mouseSensitivity = mouseSensitivity;
     }
 
-    public Settings(int[] data)
+    public Settings(float[] data)
     {
         this.sound = data[0];
         this.sfx = data[1];
         this.mouseSensitivity = data[2];
     }
 
-    public int[] toArray()
+    public float[] toArray()
     {
-        int[] data = new int[3];
+        float[] data = new float[3];
         data[0] = sound;
         data[1] = sfx;
         data[2] = mouseSensitivity;
